Add logic to group subspecies/forms/etc into species
#17

To group the different taxonranks, like subspecies, into the parent species I first extracted the species list from the data (L139-L144). Then I recalculated the scientificname based on the genus & specificEpithet. The recalculated scientificname  is then used to merge the species list with the data to overwrite the taxonkey of the respective species (L147 - L148).